% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter.R
\name{check_arguments}
\alias{check_arguments}
\title{Check arguments against type signature}
\usage{
check_arguments(fn_name, args, type_info, call_node, source_expression)
}
\arguments{
\item{fn_name}{Function name}

\item{args}{List of arguments}

\item{type_info}{Type information for function}

\item{call_node}{XML node of the call}

\item{source_expression}{Source expression}
}
\value{
Lint object or NULL
}
\description{
Check arguments against type signature
}
\keyword{internal}
