% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter-check.R
\name{check_function_calls}
\alias{check_function_calls}
\title{Check function calls against type signatures}
\usage{
check_function_calls(
  xml,
  all_types,
  var_context,
  source_expression,
  strict = FALSE
)
}
\arguments{
\item{xml}{XML parsed content}

\item{all_types}{List of all type information}

\item{var_context}{List of variable assignments}

\item{source_expression}{Source expression for creating lints}

\item{strict}{Logical, whether to warn on unknown types}
}
\value{
List of Lint objects
}
\description{
Check function calls against type signatures
}
\keyword{internal}
