% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter-check.R
\name{check_mode_annotations}
\alias{check_mode_annotations}
\title{Check for missing type annotations based on mode}
\usage{
check_mode_annotations(
  fn_assign_node,
  type_info,
  source_expression,
  mode,
  comments
)
}
\arguments{
\item{fn_assign_node}{XML node of function assignment}

\item{type_info}{Type information extracted from comments (can be NULL)}

\item{source_expression}{Source expression}

\item{mode}{Type checking mode ("lenient", "exported", "strict")}

\item{comments}{Accumulated roxygen comments for this function}
}
\value{
List of Lint objects
}
\description{
Check for missing type annotations based on mode
}
\keyword{internal}
