% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter-check.R
\name{check_strict_mode_annotations}
\alias{check_strict_mode_annotations}
\title{Check for missing type annotations in strict mode}
\usage{
check_strict_mode_annotations(fn_assign_node, type_info, source_expression)
}
\arguments{
\item{fn_assign_node}{XML node of function assignment}

\item{type_info}{Type information extracted from comments (can be NULL)}

\item{source_expression}{Source expression}
}
\value{
List of Lint objects
}
\description{
Check for missing type annotations in strict mode
}
\keyword{internal}
