% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter-check.R
\name{check_strict_mode_annotations_impl}
\alias{check_strict_mode_annotations_impl}
\title{Implementation of annotation checking (internal)}
\usage{
check_strict_mode_annotations_impl(
  fn_assign_node,
  type_info,
  source_expression,
  mode
)
}
\arguments{
\item{fn_assign_node}{XML node of function assignment}

\item{type_info}{Type information extracted from comments (can be NULL)}

\item{source_expression}{Source expression}

\item{mode}{Type checking mode (for error messages)}
}
\value{
List of Lint objects
}
\description{
Implementation of annotation checking (internal)
}
\keyword{internal}
