% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter-box.R
\name{collect_path_components}
\alias{collect_path_components}
\title{Collect path components from slash-separated expression}
\usage{
collect_path_components(expr_node)
}
\arguments{
\item{expr_node}{XML node to traverse}
}
\value{
Character vector of path components
}
\description{
Recursively traverse the expression tree to find path components.
Stops at brackets - doesn't include bracket contents in path.
}
\keyword{internal}
