% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-exports.R
\name{exported_item}
\alias{exported_item}
\title{Base Class for Exported Items}
\usage{
exported_item(name = character(0), export_type = character(0))
}
\description{
Abstract base class for all package exports. Provides common
structure for functions, classes, data objects, and constants.
}
\section{Fields}{

\describe{
\item{\code{name}}{Character\link{1} - Export name}

\item{\code{export_type}}{Character\link{1} - Type: "function", "class", "data"}
}}

\keyword{internal}
