% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter.R
\name{extract_arguments}
\alias{extract_arguments}
\title{Extract arguments from function call}
\usage{
extract_arguments(call_node)
}
\arguments{
\item{call_node}{XML node of function call}
}
\value{
List of argument information
}
\description{
Extract arguments from function call
}
\keyword{internal}
