% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter-box.R
\name{extract_box_use_arguments}
\alias{extract_box_use_arguments}
\title{Extract argument expressions from box::use() call}
\usage{
extract_box_use_arguments(call_expr)
}
\arguments{
\item{call_expr}{XML node of the box::use() call expression}
}
\value{
List of lists, each containing nodes for one argument
}
\description{
For box::use(), arguments can be:
\itemize{
\item Simple: mod/math (just an expr)
\item Aliased: m = mod/math (SYMBOL_SUB, EQ_SUB, expr)
\item Multiple: separated by OP-COMMA
}
}
\details{
We need to group nodes between commas as single argument units.
}
\keyword{internal}
