% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter.R
\name{extract_local_types}
\alias{extract_local_types}
\title{Extract type signatures from local roxygen comments}
\usage{
extract_local_types(source_expression)
}
\arguments{
\item{source_expression}{Source expression from lintr}
}
\value{
List of type information per function
}
\description{
Extract type signatures from local roxygen comments
}
\keyword{internal}
