% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet-types.R
\name{extract_s7_properties}
\alias{extract_s7_properties}
\title{Extract properties from S7::new_class call}
\usage{
extract_s7_properties(class_expr, env)
}
\arguments{
\item{class_expr}{Expression containing S7::new_class call}

\item{env}{Environment}
}
\value{
Named list of param_type objects or NULL
}
\description{
Extract properties from S7::new_class call
}
\keyword{internal}
