% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter-box.R
\name{filter_exported_functions}
\alias{filter_exported_functions}
\title{Filter function types to only include exported functions}
\usage{
filter_exported_functions(module_file, all_types)
}
\arguments{
\item{module_file}{Path to module .r/.R file}

\item{all_types}{Named list of all function types}
}
\value{
Named list of only exported function types
}
\description{
Checks which functions have #' @export directive in their documentation.
This is used to determine which functions are accessible from the module.
}
\keyword{internal}
