% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter.R
\name{find_loaded_packages}
\alias{find_loaded_packages}
\title{Find loaded packages in source}
\usage{
find_loaded_packages(source_expression)
}
\arguments{
\item{source_expression}{Source expression from lintr}
}
\value{
Character vector of package names
}
\description{
Find loaded packages in source
}
\keyword{internal}
