% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter-box.R
\name{get_cached_module_types}
\alias{get_cached_module_types}
\title{Get cached module types with mtime-based invalidation}
\usage{
get_cached_module_types(module_file)
}
\arguments{
\item{module_file}{Absolute path to module .r/.R file}
}
\value{
Named list of function types (same format as inst/types.rds)
}
\description{
Caches module type information keyed by file path + modification time.
Automatically re-parses if file has been modified.
}
\keyword{internal}
