% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet-types.R
\name{get_function_name}
\alias{get_function_name}
\title{Get function name from block}
\usage{
get_function_name(block)
}
\arguments{
\item{block}{A roxy_block object}
}
\value{
Character string or NULL
}
\description{
Get function name from block
}
\keyword{internal}
