% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet-types.R
\name{has_typed_tags}
\alias{has_typed_tags}
\title{Check if block has typed tags}
\usage{
has_typed_tags(block)
}
\arguments{
\item{block}{A roxy_block object}
}
\value{
Logical
}
\description{
Check if block has typed tags
}
\keyword{internal}
