% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter.R
\name{infer_argument_type}
\alias{infer_argument_type}
\title{Infer type of an argument from its AST node}
\usage{
infer_argument_type(arg_node)
}
\arguments{
\item{arg_node}{XML node of the argument}
}
\value{
Character string with inferred type or "unknown"
}
\description{
Infer type of an argument from its AST node
}
\keyword{internal}
