% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter-check.R
\name{is_function_top_level}
\alias{is_function_top_level}
\title{Check if a function assignment is at the top level (not nested)}
\usage{
is_function_top_level(fn_assign_node)
}
\arguments{
\item{fn_assign_node}{XML node of function assignment}
}
\value{
Logical - TRUE if function is top-level, FALSE if nested
}
\description{
Determines whether a function is defined at the top level of the file
or nested inside another function. This is used to filter out false
positives in exported mode, where nested/anonymous functions inside
closures should not be required to have type annotations.
}
\keyword{internal}
