% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-types.R
\name{is_s7_base_type}
\alias{is_s7_base_type}
\title{Check if type is an S7 base type}
\usage{
is_s7_base_type(type_string)
}
\arguments{
\item{type_string}{Type name to check}
}
\value{
Logical
}
\description{
Check if type is an S7 base type
}
\keyword{internal}
