% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet-types.R
\name{is_s7_class_definition}
\alias{is_s7_class_definition}
\title{Check if block represents an S7 class definition}
\usage{
is_s7_class_definition(block)
}
\arguments{
\item{block}{A roxy_block object}
}
\value{
Logical
}
\description{
Check if block represents an S7 class definition
}
\keyword{internal}
