% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-types.R
\name{is_union_type}
\alias{is_union_type}
\title{Check if type is a union type}
\usage{
is_union_type(type_string)
}
\arguments{
\item{type_string}{Type specification string}
}
\value{
Logical
}
\description{
Check if type is a union type
}
\examples{
is_union_type("character | NULL")
is_union_type("numeric")
}
