% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter-box.R
\name{parse_box_import_arg}
\alias{parse_box_import_arg}
\title{Parse a single box::use() import argument}
\usage{
parse_box_import_arg(arg_nodes, line)
}
\arguments{
\item{arg_nodes}{List of XML nodes making up one argument}

\item{line}{Line number of the box::use() call}
}
\value{
List with import information or NULL
}
\description{
Parse a single box::use() import argument
}
\keyword{internal}
