% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-types.R
\name{parse_function_type}
\alias{parse_function_type}
\title{Parse function type signature}
\usage{
parse_function_type(type_string)
}
\arguments{
\item{type_string}{Function signature string}
}
\value{
A list with function type information
}
\description{
Parse function type signature
}
\keyword{internal}
