% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-parser.R
\name{parse_primary_type}
\alias{parse_primary_type}
\title{Parse primary type (qualified identifier with optional element type and length)}
\usage{
parse_primary_type(parser)
}
\description{
Parse primary type (qualified identifier with optional element type and length)
}
\keyword{internal}
