% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter.R
\name{process_comment_block}
\alias{process_comment_block}
\title{Process a single comment block}
\usage{
process_comment_block(block, parsed)
}
\arguments{
\item{block}{List with lines and text}

\item{parsed}{Full parsed content}
}
\value{
List with fn_name and types
}
\description{
Process a single comment block
}
\keyword{internal}
