% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdoc.R
\name{rd}
\alias{rd}
\alias{rd_details}
\alias{rd_arguments}
\alias{rd_usage}
\alias{rd_example}
\alias{rd_examples}
\title{R docs}
\usage{
rd(topic, by_section = TRUE, include_header = TRUE,
  options = rdoc_options(), package = NULL, lib.loc = NULL)

rd_details(topic, by_section = TRUE, include_header = TRUE,
  options = rdoc_options(), package = NULL, lib.loc = NULL)

rd_arguments(topic, by_section = TRUE, include_header = TRUE,
  options = rdoc_options(), package = NULL, lib.loc = NULL)

rd_usage(topic, by_section = TRUE, include_header = TRUE,
  options = rdoc_options(), package = NULL, lib.loc = NULL)

rd_example(topic, by_section = TRUE, include_header = TRUE,
  options = rdoc_options(), package = NULL, lib.loc = NULL)
}
\arguments{
\item{topic}{\code{character(1)}, help topic}

\item{by_section}{\code{logical(1)}, should docs be shown section by section or all
at once? Defaults to TRUE.}

\item{include_header}{\code{logical(1)}, should a package header be shown with documentation? Defaults to TRUE.}

\item{options}{named list. Defaults to \code{rdoc_options()}}

\item{package}{\code{character(1)}, package of help topic. Defaults to NULL.}

\item{lib.loc}{\code{character(1)}, library location of package. Defaults to null.}
}
\description{
Refer to colourised \R docs in the console
}
\examples{
\dontrun{

rd("rd", by_section = FALSE)
rd(grepl)
rd_example("min")
rd_usage(substr)

}
}
\concept{rd access}
