% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdoc.R
\name{rd_details}
\alias{rd_details}
\title{R docs}
\usage{
rd_details(topic, by_section = TRUE, options = rdoc_options(),
  package = NULL, lib.loc = NULL)
}
\arguments{
\item{topic}{\code{character(1)}, help topic}

\item{by_section}{\code{logical(1)}, should docs be shown section by section or all
at once? Defaults to TRUE.}

\item{package}{\code{character(1)}, package of help topic. Defaults to NULL.}

\item{lib.loc}{\code{character(1)}, library location of package. Defaults to null.}
}
\description{
Refer to colourised R docs in the console
}
\examples{
\dontrun{

rd("rd", by_section = FALSE)
rd_example("min")

}
}
\seealso{
Other rd access: \code{\link{rd}}
}
\concept{rd access}
