% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdoc.R
\name{rd_question}
\alias{rd_question}
\title{rd ?}
\usage{
rd_question(type, topic)
}
\description{
rd ?
}
\details{
The manner in which content is outputted depends on where it is
  being used. If used from a terminal the contents will be outputted via
  \code{file.show}. If used from a console the output will be printed to
  screens by sections unless this has been disabled via
  \code{options(rdoc.by_section = FALSE)}. In this case pressing any key will
  print out the next section. Exiting out early is possible with the 'q' key.
  Color support is checked via \code{crayon::has_color()}.
}
