% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdoc-package.R
\docType{package}
\name{rdoc-package}
\alias{rdoc}
\alias{rdoc-package}
\title{rdoc: Type Annotations and Static Checking for R Functions}
\description{
Provides JSDoc-style type annotations for R functions using custom roxygen2 tags. Enables static type checking through lintr integration, allowing IDEs to catch type errors before runtime.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/calderonsamuel/rdoc}
  \item \url{https://calderonsamuel.github.io/rdoc/}
  \item Report bugs at \url{https://github.com/calderonsamuel/rdoc/issues}
}

}
\author{
\strong{Maintainer}: Samuel Calderon \email{samuel.calderon@uarm.pe}

}
\keyword{internal}
