% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdoc.R
\name{rdoc}
\alias{rdoc}
\title{Colourised R documentation}
\usage{
rdoc()
}
\description{
rdoc replacements for \code{?} and \code{help}
}
\details{
Calling \code{rdoc()} will override \code{utils::`?`} with \code{rd_question}
and \code{utils::help} with \code{rd}
These replacements can be unset by calling \code{rdoc()} once more.
}
\examples{
\dontrun{

rdoc::rdoc()
?help
rdoc::rdoc()
?help
}

}
