% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdoc.R
\name{rdoc_options}
\alias{rdoc_options}
\title{R doc options}
\usage{
rdoc_options(pkg_header = cli::rule, title = crayon::bold,
  section_titles = crayon::underline, arg_params = crayon::underline,
  code_style = prettycode::default_style(),
  item_bullet = cli::symbol$bullet)
}
\arguments{
\item{pkg_header}{package header information. Defaults to \code{cli::rule}}

\item{title}{topic title. Defaults to \code{crayon::bold}}

\item{section_titles}{section titles Defaults to \code{crayon::underline}}

\item{arg_params}{argument paramets. Defaults to \code{crayon::underline}}

\item{code_style}{code sections and inline code. Defaults to \code{prettycode::default_style()}}

\item{item_bullet}{item bullet points. Defaults to \code{cli::symbol$bullet}}
}
\description{
Set output formatting styles
}
