% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdoc.R
\name{rdoc_rd}
\alias{rdoc_rd}
\title{Colourised \R documentation}
\usage{
rdoc_rd(path)
}
\arguments{
\item{path}{character(1), the path to an .Rd file}
}
\description{
Refer directly to .Rd file contents as terminal output.
}
\details{
The manner in which content is outputted depends on where it is
  being used. If used from a terminal the contents will be outputted via
  \code{file.show}. Otherwise the output will be printed by sections unless
  this has been disabled via \code{options(rdoc.by_section = FALSE)}. When
  printing by section pressing any key will print out the next section.
  Exiting out early is possible via 'q' keypress. Color support is checked
  via \code{crayon::has_color()}. The output styles can be modified via
  \code{rdoc_style()} and passing this to the option:
  \code{options(rdoc.style = rdoc_style())}
}
