% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{rdoc_style}
\alias{rdoc_style}
\title{Rdoc style options}
\usage{
rdoc_style(title = crayon::bold, section_titles = crayon::underline,
  arg_params = crayon::underline,
  code_style = prettycode::default_style(),
  item_bullet = cli::symbol$bullet)
}
\arguments{
\item{title}{topic title. Defaults to \code{crayon::bold}}

\item{section_titles}{section titles Defaults to \code{crayon::underline}}

\item{arg_params}{argument parameters. Defaults to \code{crayon::underline}}

\item{code_style}{code sections and inline code. Defaults to
\code{prettycode::default_style()}}

\item{item_bullet}{item bullet points. Defaults to \code{cli::symbol$bullet}}
}
\value{
A list
}
\description{
Set Rd output formatting styles to be supplied to \code{options(rdoc.style =
rdoc_style())}
}
\details{
all inputs should be functions with the exception of
  \code{code_style} which should be a list of functions and
  \code{item_bullet} which should be a \code{character(1)}.
}
