% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdoc.R
\name{rdoc_text}
\alias{rdoc_text}
\title{Colourised \R documentation}
\usage{
rdoc_text(topic, package = NULL, lib.loc = NULL)
}
\arguments{
\item{topic}{\code{character(1)}, help topic}

\item{package}{\code{character(1)}, package of help topic. Defaults to
\code{NULL}}

\item{lib.loc}{\code{character(1)}, library location of package. Defaults to
\code{NULL}}
}
\value{
character
}
\description{
Access the underlying rdoc text lines
}
\examples{
txt <- rdoc::rdoc_text('min')
}
