% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{rdoc_text_formats}
\alias{rdoc_text_formats}
\title{Rdoc text formatting}
\usage{
rdoc_text_formats(r_logo = crayon::combine_styles("bold", "blue"),
  pkg = function(x) paste0("{", x, "}"),
  inline_code = prettycode::default_style(), italic = crayon::italic,
  bold = crayon::bold, squotes = NULL, email = NULL, url = NULL,
  href = NULL)
}
\arguments{
\item{r_logo}{\R symbol}

\item{pkg}{\pkg{pkg} references}

\item{inline_code}{inline code}

\item{italic}{italic font}

\item{bold}{bold font}

\item{squotes}{single quotes}

\item{email}{email}

\item{url}{url}

\item{href}{href}
}
\description{
Set Rd text format defaults to be supplied to \code{options(rdoc.text_formats
= rdoc_text_formats())}
}
\details{
All inputs must either be a function (exceptionally as a list of
style functions for `inline_code`) or NULL. Functions must take and return a
vector: \code{character(1)} If set as the latter, the relevant text will be
displayed using the defaults of \code{tools::Rd2txt}
}
