% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-metadata.R
\name{return_type}
\alias{return_type}
\title{Return Type Information}
\usage{
return_type(type = character(0), description = character(0))
}
\arguments{
\item{type}{Character string with the type specification}

\item{description}{Character string describing the return value}
}
\description{
Represents type information for a function's return value.
}
\keyword{internal}
