% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet-types.R
\name{roclet_output.roclet_types}
\alias{roclet_output.roclet_types}
\title{Output type metadata to file}
\usage{
\method{roclet_output}{roclet_types}(x, results, base_path, ..., is_first = TRUE)
}
\arguments{
\item{x}{The roclet}

\item{results}{Results from roclet_process}

\item{base_path}{Base path of the package}

\item{...}{Additional arguments}

\item{is_first}{Whether this is the first roclet run}
}
\value{
Path to generated file
}
\description{
Output type metadata to file
}
