% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{roclet_tags.roclet_rd}
\alias{roclet_tags.roclet_rd}
\title{Register tags with rd roclet}
\usage{
\method{roclet_tags}{roclet_rd}(x)
}
\arguments{
\item{x}{The roclet}
}
\value{
List of tag configurations
}
\description{
Tells roxygen2's rd roclet about our custom tags
}
\keyword{internal}
