% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{roxy_tag_rd.roxy_tag_typedParam}
\alias{roxy_tag_rd.roxy_tag_typedParam}
\title{Convert @typedParam to rd_section}
\usage{
\method{roxy_tag_rd}{roxy_tag_typedParam}(x, base_path, env)
}
\arguments{
\item{x}{A roxy_tag_typedParam object}

\item{base_path}{Base path}

\item{env}{Environment}
}
\value{
An rd_section object
}
\description{
Convert @typedParam to rd_section
}
\keyword{internal}
