% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet-types.R
\name{s7_class_expr_to_string}
\alias{s7_class_expr_to_string}
\title{Convert S7 class expression to string}
\usage{
s7_class_expr_to_string(expr)
}
\arguments{
\item{expr}{Expression representing an S7 class type}
}
\value{
Character string
}
\description{
Convert S7 class expression to string
}
\keyword{internal}
