% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tag_parse_typed_param}
\alias{tag_parse_typed_param}
\title{Internal parser for typed-param}
\usage{
tag_parse_typed_param(x)
}
\arguments{
\item{x}{A roxy_tag object}
}
\value{
A list with param, type, and description
}
\description{
Parses the format: "param_name {type} description"
}
\keyword{internal}
