% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-parser.R
\name{validate_null_position}
\alias{validate_null_position}
\title{Validate NULL position in union types}
\usage{
validate_null_position(types)
}
\arguments{
\item{types}{List of type AST nodes from a union}
}
\description{
Enforces rdoc's opinionated rule: NULL must come first in unions.
This matches S7's convention where union order determines default value.
}
\keyword{internal}
