% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter-validate.R
\name{validate_return_type}
\alias{validate_return_type}
\title{Validate function return type matches declaration}
\usage{
validate_return_type(fn_assign_node, declared_type, source_expression)
}
\arguments{
\item{fn_assign_node}{XML node of function assignment}

\item{declared_type}{Declared return type from @typedReturn}

\item{source_expression}{Source expression for creating lints}
}
\value{
List of Lint objects (empty if validation passes or cannot be performed)
}
\description{
Validate function return type matches declaration
}
\keyword{internal}
