% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-types.R
\name{validate_type_spec}
\alias{validate_type_spec}
\title{Validate a type specification}
\usage{
validate_type_spec(type_string)
}
\arguments{
\item{type_string}{Type specification to validate}
}
\value{
Logical indicating if valid
}
\description{
Checks if a type specification is valid
}
\keyword{internal}
