% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-metadata.R
\docType{data}
\name{xml_node}
\alias{xml_node}
\title{S7 Wrapper for xml2::xml_node}
\format{
An object of class \code{S7_S3_class} of length 3.
}
\usage{
xml_node
}
\description{
Wraps the xml_node S3 class from xml2 package for type-safe usage in S7 classes.
This allows us to specify xml_node as a property type instead of using class_any.
}
\details{
Used throughout rdoc for:
\itemize{
\item Argument nodes in call_argument
\item Variable assignment value nodes
\item Return expression nodes
\item Any XML AST node from xmlparsedata
}
}
\keyword{internal}
