% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter.R
\name{check_single_call}
\alias{check_single_call}
\title{Check a single function call}
\usage{
check_single_call(call_node, all_types, source_expression)
}
\arguments{
\item{call_node}{XML node of the function call}

\item{all_types}{List of type information}

\item{source_expression}{Source expression}
}
\value{
Lint object or NULL
}
\description{
Check a single function call
}
\keyword{internal}
