% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter-check.R
\name{check_single_call}
\alias{check_single_call}
\title{Check a single function call}
\usage{
check_single_call(
  call_node,
  all_types,
  var_context,
  source_expression,
  strict = FALSE
)
}
\arguments{
\item{call_node}{XML node of the function call}

\item{all_types}{List of type information}

\item{var_context}{List of variable assignments}

\item{source_expression}{Source expression}

\item{strict}{Logical, whether to warn on unknown types}
}
\value{
List of Lint objects
}
\description{
Check a single function call
}
\keyword{internal}
