% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter-inference.R
\name{extract_arguments}
\alias{extract_arguments}
\title{Extract arguments from function call}
\usage{
extract_arguments(call_node, var_context = NULL, current_line = NULL)
}
\arguments{
\item{call_node}{XML node of function call}

\item{var_context}{Optional list of variable assignments for type inference}

\item{current_line}{Optional line number for variable lookup}
}
\value{
List of argument information
}
\description{
Extract arguments from function call
}
\keyword{internal}
