% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter-box.R
\name{extract_module_types}
\alias{extract_module_types}
\title{Extract type annotations from a box module file}
\usage{
extract_module_types(module_file)
}
\arguments{
\item{module_file}{Path to module .r/.R file}
}
\value{
Named list: function_name -> list(params = list(...), return = "type")
}
\description{
Reads a .r/.R file and extracts all type annotations from functions
marked with #' @export. Returns type information in the same format
as inst/types.rds for consistency with package type loading.
}
\examples{
\dontrun{
# Module file:
# #' @typedParam x {class_numeric} value
# #' @typedReturn {class_numeric} result
# #' @export
# double <- function(x) x * 2

types <- extract_module_types("math.r")
# Returns: list(double = list(params = list(x = "numeric"), return = "numeric"))
}
}
\keyword{internal}
