% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet-types.R
\name{extract_type_info_from_block}
\alias{extract_type_info_from_block}
\title{Extract type information from a roxygen block}
\usage{
extract_type_info_from_block(block)
}
\arguments{
\item{block}{A roxy_block object}
}
\value{
A list with params and return type info
}
\description{
Extract type information from a roxygen block
}
\keyword{internal}
