% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter-extract.R
\name{extract_types_from_comment_lines}
\alias{extract_types_from_comment_lines}
\title{Extract types from a vector of comment lines}
\usage{
extract_types_from_comment_lines(comments)
}
\arguments{
\item{comments}{Character vector of comment lines}
}
\value{
List with params and return type info
}
\description{
Extract types from a vector of comment lines
}
\keyword{internal}
