% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter.R
\name{extract_types_from_comments}
\alias{extract_types_from_comments}
\title{Extract type information from comment blocks}
\usage{
extract_types_from_comments(comments, parsed)
}
\arguments{
\item{comments}{Data frame of comment lines}

\item{parsed}{Full parsed content}
}
\value{
List of type information
}
\description{
Extract type information from comment blocks
}
\keyword{internal}
