% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter.R
\name{extract_types_from_xml}
\alias{extract_types_from_xml}
\title{Extract types from XML}
\usage{
extract_types_from_xml(xml)
}
\arguments{
\item{xml}{XML parsed content}
}
\value{
List of type information
}
\description{
Extract types from XML
}
\keyword{internal}
