% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter-inference.R
\name{infer_argument_type}
\alias{infer_argument_type}
\title{Infer type of an argument from its AST node}
\usage{
infer_argument_type(arg_node, var_context = NULL, current_line = NULL)
}
\arguments{
\item{arg_node}{XML node of the argument}

\item{var_context}{Optional list of variable assignments for lookup}

\item{current_line}{Optional line number where argument is used (for variable lookup)}
}
\value{
Character string with inferred type or "unknown"
}
\description{
Infer type of an argument from its AST node
}
\keyword{internal}
