% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter-validate.R
\name{infer_function_return_type}
\alias{infer_function_return_type}
\title{Infer actual return type from function body}
\usage{
infer_function_return_type(body_node)
}
\arguments{
\item{body_node}{XML node of function body}
}
\value{
Character string with inferred type or "unknown"/"complex"
}
\description{
Infer actual return type from function body
}
\keyword{internal}
