% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-types.R
\name{is_base_type}
\alias{is_base_type}
\title{Check if type is a base R type}
\usage{
is_base_type(type_string)
}
\arguments{
\item{type_string}{Type to check}
}
\value{
Logical
}
\description{
Check if type is a base R type
}
\keyword{internal}
