% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter.R
\name{load_package_types}
\alias{load_package_types}
\title{Load type metadata from packages}
\usage{
load_package_types(packages)
}
\arguments{
\item{packages}{Character vector of package names}
}
\value{
List of type information from packages
}
\description{
Load type metadata from packages
}
\keyword{internal}
