% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter.R
\name{parse_typed_param_text}
\alias{parse_typed_param_text}
\title{Parse typed param text}
\usage{
parse_typed_param_text(text)
}
\arguments{
\item{text}{Text after @typedParam}
}
\value{
List with param, type, description
}
\description{
Parse typed param text
}
\keyword{internal}
