% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter-extract.R
\name{parse_typed_return_text}
\alias{parse_typed_return_text}
\title{Parse typed return text}
\usage{
parse_typed_return_text(text)
}
\arguments{
\item{text}{Text after @typedReturn}
}
\value{
List with type, description
}
\description{
Parse typed return text
}
\keyword{internal}
