% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdoc.R
\name{rd}
\alias{rd}
\alias{rd_details}
\alias{rd_arguments}
\alias{rd_usage}
\alias{rd_example}
\alias{rd_examples}
\title{Colourised R documentation}
\usage{
rd(topic, package = NULL, lib.loc = NULL)

rd_details(topic, package = NULL, lib.loc = NULL)

rd_arguments(topic, package = NULL, lib.loc = NULL)

rd_usage(topic, package = NULL, lib.loc = NULL)

rd_example(topic, package = NULL, lib.loc = NULL)
}
\arguments{
\item{topic}{\code{character(1)}, help topic}

\item{package}{\code{character(1)}, package of help topic. Defaults to
\code{NULL}}

\item{lib.loc}{\code{character(1)}, library location of package. Defaults to
\code{NULL}}
}
\description{
Refer to colourised \R docs as terminal/console output. Provides a
replacement for \code{help}. A number of common section accessors are also
provided.
}
\details{
The manner in which content is outputted depends on where it is
  being used. If used from a terminal the contents will be outputted via
  \code{file.show}. If used from a console the output will be printed to
  screens by sections unless this has been disabled via
  \code{options(rdoc.by_section = FALSE)}. In this case pressing any key will
  print out the next section. Exiting out early is possible with the 'q' key.
  Color support is checked via \code{crayon::has_color()}.
}
\examples{
\dontrun{

rd("rd")
rd(grepl)
rd_example("min")
rd_usage(substr)

}
}
\concept{rd access}
