% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdoc.R
\name{rdoc}
\alias{rdoc}
\title{rdoc
rdoc-style \code{?} replacement for \code{utils::`?`}}
\usage{
rdoc()
}
\description{
rdoc
rdoc-style \code{?} replacement for \code{utils::`?`}
}
\details{
Calling \code{rdoc()} will override \code{utils::`?`}
}
