% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdoc.R
\name{rdoc}
\alias{rdoc}
\alias{rd}
\title{R docs}
\usage{
rdoc(topic, by_section = TRUE, options = rdoc_options(),
  package = NULL, lib.loc = NULL)
}
\arguments{
\item{topic}{character(1), help topic}

\item{by_section}{logical(1), should docs be shown section by section or all
at once? Defaults to TRUE.}

\item{package}{character(1), package of help topic. Defaults to NULL.}

\item{lib.loc}{character(1), library location of package. Defaults to null.}
}
\description{
Refer to colourised R docs in the console
}
\examples{
\dontrun{

rd("min", by_section = FALSE)

}
}
