% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{package}
\name{rdoc_options}
\alias{rdoc_options}
\alias{rdoc_options-package}
\title{rdoc options}
\description{
Rdoc uses the following \code{options()} to set output:

  \itemize{ \item `rdoc.by_section`: logical, output printed by sections for
  console sessions. Default is TRUE.

  \item `rdoc.header`: logical, whether package header is included in the output.
  Default is TRUE.

  \item `rdoc.text_formats`: see \code{\link{rdoc_text_formats}}

  \item `rdoc.style`: see \code{\link{rdoc_style}}

  \item `rdoc.item_bullet`: Symbol to use for itemized lists.
  Default is \code{paste0(cli::symbol$circle, " ")}
}
}
