% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdoc.R
\name{rdoc_options}
\alias{rdoc_options}
\title{R doc options}
\usage{
rdoc_options(pkg_header = cli::rule, title = crayon::bold,
  section_titles = crayon::underline, arg_params = crayon::underline,
  code = prettycode::highlight, item_bullet = cli::symbol$bullet)
}
\description{
Set output styles
}
