% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdoc.R
\name{rdoc_rd}
\alias{rdoc_rd}
\title{Colourised \R documentation}
\usage{
rdoc_rd(path)
}
\arguments{
\item{path}{character(1), the path to an .Rd file}
}
\description{
Refer directly to .Rd file contents as terminal output.
}
\details{
The manner in which content is outputted depends on where it is
  being used. If used from a terminal the contents will be outputted via
  \code{file.show}. Otherwise the output will be printed interactively by
  sections unless this has been disabled via \code{options(rdoc.by_section =
  FALSE)}. When printing by section pressing the <enter> keypress will show
  the next section. Any other keypress will exit the interaction without
  printing any remaining sections. Color support is checked via
  \code{crayon::has_color()}. Customising rdoc output is possible via
  `options`, see \code{\link{rdoc_options}} for more details.
}
\examples{
\donttest{
d <- system.file('extdata/rdoc_test.Rd', package = "rdoc")
rdoc::rdoc_rd(d)
}

}
