% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{rdoc_text_formats}
\alias{rdoc_text_formats}
\title{Rdoc text formatting}
\usage{
rdoc_text_formats(r_logo = R_logo, pkg = function(x) paste0("{", x,
  "}"), italic = crayon::italic, bold = crayon::bold, squotes = NULL,
  email = NULL, url = NULL, href = NULL)
}
\arguments{
\item{r_logo}{\R symbol}

\item{pkg}{\pkg{pkg} formatting}

\item{italic}{italic font}

\item{bold}{bold font}

\item{squotes}{single quotes}

\item{email}{email}

\item{url}{= NULL}

\item{href}{= NULL}
}
\description{
Set Rd text format defaults to be supplied to \code{options(rdoc.text_formats =
rdoc_text_formats())}
}
\details{
All inputs must either be a function or NULL. If set as the latter, the relevant text
will be displayed using the defaults of \code{tools::Rd2txt}
}
