% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet-types.R
\name{roclet_process.roclet_types}
\alias{roclet_process.roclet_types}
\title{Process blocks to extract type information}
\usage{
roclet_process.roclet_types(x, blocks, env, base_path)
}
\arguments{
\item{x}{The roclet}

\item{blocks}{List of roxy_block objects}

\item{env}{Environment}

\item{base_path}{Base path of the package}
}
\value{
List of type information per function
}
\description{
Process blocks to extract type information
}
