% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{roxy_tag_parse.roxy_tag_typedParam}
\alias{roxy_tag_parse.roxy_tag_typedParam}
\title{Parse @typedParam tag}
\usage{
\method{roxy_tag_parse}{roxy_tag_typedParam}(x)
}
\arguments{
\item{x}{A roxy_tag object}
}
\value{
A roxy_tag object with param class for roxygen2 compatibility
}
\description{
This tag works as both @param (for roxygen2 documentation) and provides
type information for static type checking. Format: name \{type\} description
}
