% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{roxy_tag_parse.roxy_tag_typedParam}
\alias{roxy_tag_parse.roxy_tag_typedParam}
\title{Parse @typedParam tag}
\usage{
\method{roxy_tag_parse}{roxy_tag_typedParam}(x)
}
\arguments{
\item{x}{A roxy_tag object}
}
\value{
A list with param, type, and description
}
\description{
Parse @typedParam tag
}
