% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{roxy_tag_parse.roxy_tag_typedReturn}
\alias{roxy_tag_parse.roxy_tag_typedReturn}
\title{Parse @typedReturn tag}
\usage{
\method{roxy_tag_parse}{roxy_tag_typedReturn}(x)
}
\arguments{
\item{x}{A roxy_tag object}
}
\value{
A list with type and description
}
\description{
Parse @typedReturn tag
}
