% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-types.R
\name{s7_class_to_type_string}
\alias{s7_class_to_type_string}
\title{Convert S7 class object to type string}
\usage{
s7_class_to_type_string(s7_class)
}
\arguments{
\item{s7_class}{S7 class object}
}
\value{
Full S7 class name (e.g., "class_integer")
}
\description{
Convert S7 class object to type string
}
\keyword{internal}
