% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-types.R
\name{split_union_types}
\alias{split_union_types}
\title{Split union type into components}
\usage{
split_union_types(type_string)
}
\arguments{
\item{type_string}{Union type string}
}
\value{
Character vector of individual types
}
\description{
Split union type into components
}
\examples{
split_union_types("character | NULL")
split_union_types("numeric | integer | complex")
}
