% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tag_parse_typed_return}
\alias{tag_parse_typed_return}
\title{Internal parser for typed-return}
\usage{
tag_parse_typed_return(x)
}
\arguments{
\item{x}{A roxy_tag object}
}
\value{
A list with type and description
}
\description{
Parses the format: "{type} description"
}
\keyword{internal}
