% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter.R
\name{types_compatible}
\alias{types_compatible}
\title{Check if two types are compatible}
\usage{
types_compatible(actual, expected)
}
\arguments{
\item{actual}{Actual type}

\item{expected}{Expected type}
}
\value{
Logical
}
\description{
Check if two types are compatible
}
\keyword{internal}
